refactoring: - move offer model and mocks to shared
