+++ src/entities/Offer/index.ts
@@ -1,3 +1,1 @@
-export * from './mocks';
 export * from './ui';
-export * from './model';
