--- src/entities/Offer/mocks/index.ts
@@ -1,1 +1,0 @@
-export * from './offersMock';
