+++ src/entities/Offer/ui/OfferCard/OfferCard.type.ts
@@ -1,4 +1,4 @@
-import {TOffer} from '../../model/offer';
+import {TOffer} from '@/shared/model/offer/offer.ts';
 
 export interface IOfferCard {
   offer: TOffer;
