+++ src/shared/model/offer/index.ts
