+++ src/shared/model/offer/offer.ts
